drivers/at24cxxx: merge mtd driver with at24cxxx.c
