+++ drivers/at24cxxx/Makefile
@@ -1,5 +1,1 @@
-ifneq (,$(filter mtd_at24cxxx,$(USEMODULE)))
-  DIRS += mtd
-endif
-
 include $(RIOTBASE)/Makefile.base
