+++ drivers/at24cxxx/Makefile.include
@@ -1,4 +1,5 @@
 PSEUDOMODULES += at24c%
+PSEUDOMODULES += mtd_at24cxxx
 # handle at24cxxx being a distinct module
 NO_PSEUDOMODULES += at24cxxx
