--- drivers/at24cxxx/mtd/Makefile
@@ -1,10 +1,0 @@
-MODULE := mtd_at24cxxx
-BASE_MODULE := at24cxxx
-
-# at24cxxx_mtd files
-SRC := mtd.c
-
-# enable submodules
-SUBMODULES := 1
-
-include $(RIOTBASE)/Makefile.base
